Fixing old tabs
